pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embedded-analyzer"
version = "1.0.0"
description = "Embedded object section size analyzer - CLI & TUI"
readme = "README.md"
authors = [
    {name = "ESA Contributors"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
tui = ["textual>=0.40"]
dev = [
    "pytest>=7.4.0",
    "textual>=0.40",
]

[project.scripts]
esa = "embedded_analyzer.cli:main"

[tool.setuptools]
packages = ["embedded_analyzer"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--verbose",
    "--tb=short",
]